app/backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy>=2.0
pydantic>=2.7
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
websockets>=12.0
aiofiles>=23.0
